requirements.txt
streamlit
yfinance
numpy
pandas
matplotlib
scikit-learn
tensorflow-cpu==2.12.0
